pyproject.toml
[project]
name = "medical-qa-system"
version = "0.1.0"
description = "基于RAG的医疗知识问答系统"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "llama-index>=0.10.0",
    "llama-index-embeddings-ollama>=0.1.0",
    "llama-index-llms-ollama>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
    "python-docx>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    # API服务
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["app", "backend", "rag"]
